Add support for x-request-id header
This header allows the frontend/client to specify a custom request ID
that will be propagated back to the client in the response, as well as
any outgoing event that was generated as a result of this request.

It may allow some interesting features in the frontend, like optimistic
UI updates. It may help correlate issues and improve debugging (for
example, by knowing which request originated an unexpected event in the
frontend).

Within the backend, this is also useful for tests! One example is being
able to query every event that was emitted as a result of the request
identified by `x-request-id`.