Scope Server/Player context based on the Universe they belong to
We can't simply say there's a :server shard with shard_id=1. Because,
for both Singleplayer and Multiplayer universes, there will be a Server
with id=1, and it won't be the same server!

Therefore, we need to scope the Server shard based on which Universe
they belong to: `sp_server` or `mp_server`. Same for Player.

Note: theoretically, this scoping is not needed if the Singleplayer
server (lowercase "s", meaning I'm talking about the actual game
infrastructure) is different from the multiplayer server. In that
scenario, there would not be a conflict for Server with shard_id=1 for
both Singleplayer and Multiplayer, because they live, phisically, in a
different disk/server.