Start and finish DB transaction before delegating event to Handler
With this change, the Core.Event logic will be managing the DB
connection that is used by Handlers. If needed, each Handler can
modify the default behaviour based on the implemented callbacks.

A similar behaviour can be found in the requests: the DB transaction is
managed by the Webserver.Dispatcher, and each Endpoint handles the
request without having to worry how/when to start/close the connection.

The idea is the same: provide an easy API for event handlers, and shift
the complexity of managing the DB connections to the Core.Event logic.