Add experimental script to support "named unused variables"
Personally, I'm a big fan of prefixing unused variables with an
underscore, so we can have something like _appMsg = Foo.

It helps clarify what value is present in the variable, even if it's
unused. Elixir supports this, Elm doesn't.

In fact, Elm doesn't allow variable names to begin with an underscore.

Because I mando na porra toda, I've decided to write a script to replace
variables with double underscores when running elm-review (i.e. when
checking for unused variables).

By doing this, I can have my cake and eat it too: I get to name an
unused variable, and I can still get the benefits of NoUnused rules.

Surely, it's an unorthodox approach -- let's see how it plays out.