Initial work on Henforcer
Pretty much exactly the same pattern as old Helix (except for the
macros).

One interesting difference is that now we don't _need_ to care so much
about excessive queries, since the RTT overhead is effectively zero, but
still it makes sense to be careful and have a properly designed
authorization layer.

Better documentation and better testing is TODO. Now is not the time to
focus on Henforcer itself.