Initial work on TOP
TOP is a GenServer that exists for each in-game Server with active
processes. It is responsible for allocating resources and scheduling the
next process to be completed.

It is also the (sequential, single-threaded) point of origin for all
things process: executing a process, completing a process, or anything
in-between (say, pause/resume/retarget/renice) goes through the TOP.