Add ServerMeta
This will include a "cache" of the total resources within the server.
Actual hardware implementation is TODO, but for the time being we need
to know how many (much?) resources a server has in order to properly
schedule its processes.

In the future this table will also hold the server password. It is
supposed to have only one entry (row) per shard.