Refactor Log schema to add the concept of `direction`
Up until now, the "direction" was "embedded" into the type of the log.
This meant that, using `file_deleted` log as an example, we'd need:
`local_file_deleted`, `remote_file_deleted_gateway` and
`remote_file_deleted_endpoint`.

This will prove to be a challenge once we implement custom filtering
support in the LogScanner, plus it seems to be way more organized to
have this information stored in a separate column.