Improve SSEListener test util for less flaky behavior
Now we can specify exaclty which event SSEListener is waiting for.
Without this change, adding a new event that would be emitted *prior* to
the target event would mean updating all tests and incrementing the old
`total_expected_events` timer.

We can still support that, but it's better to just say exactly which
event you are looking for.